ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
qdrant-client>=1.7.0
groq>=0.4.0
python-docx>=0.8.11